Changed folder layout.
