+++ src/perovskite_solar_cell_database/schema_sections/cell.py
