+++ src/perovskite_solar_cell_database/schema_sections/vars.py
